BSHUB-429: Tidy up code
